Append eglot-check instead of setting disabled checkers
When `flycheck-eglot-mode` is disabled in the buffer, the checker `eglot-check`
should not take effects. Adding it to the disabled checker list is a good idea,
however, there's no guarantee that the user has chance to call
`flycheck-eglot--teardown`, I mean, to toggle `flycheck-eglot-mode` before they
kill the buffer.

As a result, `eglot-check` remains at the begining of `flycheck-checkers`, which
will be selected by flycheck next time when the user change to the same major
mode. Even worse, if the user has set `flycheck-eglot-exclusive`, the other
checkers have no chances to run. And eglot provides no linting diagnostic if
`flycheck-eglot-mode` is disabled, which is not a desired situation.

In fact, I think the buffer-local mode `flycheck-eglot-mode` should not modify
any global variables to achieve the expected states, in this case,
`flycheck-disabled-checkers` and `flycheck-checkers`.

The workaround here:
- Define a new function `flycheck-eglot--register-eglot-checker`, which register
  `elgot-check` as a valid flycheck checker. The register routine is idempotent.
- It appends `eglot-check` **in the end** of `flycheck-checkers`. As a result,
  it will not be selected as the default checker by flycheck.
- If the user has not set `flycheck-eglot-exclusive`, add the first valid
  checker for the major mode to the next checker of `eglot-checker`.
- When `flycheck-eglot-mode` is enabled, always set the buffer-local
  `flycheck-checker` to `eglot-check`. Therefore, it will always be run at
  first, and depending on `flycheck-eglot-exclusive`, flycheck may or may not select
  the next checker(s).